refactor: move custom storage and network adapter implementations to their own file
